Word change
Suggest change of word from "inappropriate" to "unwelcome" in "Inappropriate photography or recording." I believe this bullet point is meant to describe when taking a photograph (versus sharing a photograph) should not be done. That should be cases where it is "unwelcome" by the subject, not whether or not the image itself would be offensive in some way.